onlineshop/requirements.txt
loremipsum==1.0.5
selenium==3.7.0
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.5
Pillow==2.3.0
psycopg2==2.7.3.1
static3==0.7.0
gunicorn==19.6.0